refactor: create MemStore
